Fixed folder structure, builds properly now
