--- src/main/kotlin/io/github/pylonmc/pylon/core/persistence/PylonDataWriter.kt
@@ -1,8 +1,0 @@
-package io.github.pylonmc.pylon.core.persistence
-
-import org.bukkit.NamespacedKey
-import org.bukkit.persistence.PersistentDataType
-
-interface PylonDataWriter {
-    fun <P : Any?, C : Any?> set(key: NamespacedKey, type: PersistentDataType<P, C>, value: C & Any)
-}
